% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregacoes.R
\name{.mapa_municipios_interno}
\alias{.mapa_municipios_interno}
\title{Mapa TSE x IBGE interno (conexao padrao via env vars)}
\usage{
.mapa_municipios_interno(anos, uf)
}
\description{
Mapa TSE x IBGE interno (conexao padrao via env vars)
}
\keyword{internal}
