% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn.R
\name{.tse_cdn_hash_url}
\alias{.tse_cdn_hash_url}
\title{Caminho do arquivo de hash SHA-512 publicado em par com um zip}
\usage{
.tse_cdn_hash_url(url_zip)
}
\description{
Caminho do arquivo de hash SHA-512 publicado em par com um zip
}
\keyword{internal}
