% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregacoes.R
\name{.uf_para_ibge}
\alias{.uf_para_ibge}
\title{Converte sigla de UF em codigo IBGE de 2 digitos}
\usage{
.uf_para_ibge(uf)
}
\description{
Converte sigla de UF em codigo IBGE de 2 digitos
}
\keyword{internal}
