% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsebr-package.R
\docType{package}
\name{tsebr-package}
\alias{tsebr}
\alias{tsebr-package}
\title{tsebr: Acesso a Dados Eleitorais do TSE}
\description{
Cliente R para o Portal de Dados Abertos do TSE (CKAN) e a CDN de
arquivos: descoberta de datasets, download com verificacao SHA-512,
leitura dos CSVs (Latin-1, separador ponto-e-virgula) e conformacao
de series eleitorais em formato longo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DistintiveLab/tsebr}
  \item Report bugs at \url{https://github.com/DistintiveLab/tsebr/issues}
}

}
\author{
\strong{Maintainer}: DistintiveLab \email{contato@distintive.com.br}

}
\keyword{internal}
